Uploaded the Mobile-learning Java EE MVC project from netbeans
This is an MVC based Java EE project which includes java, JSP, servlet, html, css and javascript. This project is done for creating a mobile learning system wherein the employees of the organization or the company can learn about their departments and other information units uploaded by the organization. The organization can monitor the employee learning rate and pull reports based on that.